requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
tqdm>=4.66.0
pyrogram
tgcrypto
python-dotenv
